+++ src/Classifier/MarkdownClassifierProvider.cs
@@ -1,25 +1,16 @@
-﻿using System;
-using System.ComponentModel.Composition;
+﻿using System.ComponentModel.Composition;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Classification;
-using Microsoft.VisualStudio.Text.Tagging;
 using Microsoft.VisualStudio.Utilities;
 
 namespace MarkdownEditor
 {
     [Export(typeof(IClassifierProvider))]
-    [Export(typeof(ITaggerProvider))]
-    [TagType(typeof(IOutliningRegionTag))]
     [ContentType(MarkdownLanguage.LanguageName)]
-    internal class MarkdownClassifierProvider : IClassifierProvider, ITaggerProvider
+    internal class MarkdownClassifierProvider : IClassifierProvider
     {
         [Import]
         private IClassificationTypeRegistryService classificationRegistry { get; set; }
-
-        public ITagger<T> CreateTagger<T>(ITextBuffer buffer) where T : ITag
-        {
-            return buffer.Properties.GetOrCreateSingletonProperty(() => new MarkdownClassifier(buffer, classificationRegistry)) as ITagger<T>;
-        }
 
         public IClassifier GetClassifier(ITextBuffer buffer)
         {