+++ src/Outlining/MarkdownOutliningProvider.cs
@@ -1,0 +1,21 @@
+using System;
+using System.ComponentModel.Composition;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Classification;
+using Microsoft.VisualStudio.Text.Tagging;
+using Microsoft.VisualStudio.Utilities;
+
+
+namespace MarkdownEditor.Outlining
+{
+    [Export(typeof(ITaggerProvider))]
+    [TagType(typeof(IOutliningRegionTag))]
+    [ContentType(MarkdownLanguage.LanguageName)]
+    public class MarkdownOutliningProvider : ITaggerProvider
+    {
+        public ITagger<T> CreateTagger<T>(ITextBuffer buffer) where T : ITag
+        {
+            return buffer.Properties.GetOrCreateSingletonProperty(() => new MarkdownOutliningTagger(buffer)) as ITagger<T>;
+        }
+    }
+}