+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Roadmap
 
 - [x] Item template for Markdown file
+- [x] Outlining per heading (#21)
 - [ ] Buttons to switch between Split/Source/Rendered mode
 - [ ] Show/hide the preview window when Options change
 - [ ] Format document/selection (#12)